Edit

+++ src/main/java/faang/school/godbless/BJS2_23905/CityWorker.java
@@ -30,6 +30,6 @@ public class CityWorker implements Runnable {
     }
 
     public long getJourneyDistance() {
-        return (long) city.getDistance();
+        return city.getDistance();
     }
 }
